Moved existing hooks into the same folder and suppressed STDOUT output
