+++ python/train.py
@@ -10,11 +10,11 @@ import tensorflow as tf
 from datetime import datetime
 
 from python.params import FLAGS, get_parameters
-from python.util import storage, tf_contrib
+from python.util import storage
 import python.model as model
 from python.evaluate import evaluate
 
-from python.util.hooks import GPUStatisticsHook
+from python.util.hooks import GPUStatisticsHook, LoggerHook
 
 
 # General TensorFlow settings and setup.
@@ -107,9 +107,10 @@ def train(epoch):
 
         # GPU statistics hook.
         gpu_stats_hook = GPUStatisticsHook(
-            every_n_steps=FLAGS.log_frequency * 40,
+            every_n_steps=FLAGS.log_frequency,
             stats=['mem_util', 'gpu_util'],
-            summary_writer=file_writer
+            summary_writer=file_writer,
+            suppress_stdout=False
         )
 
         # Stop after steps hook.
@@ -132,7 +133,7 @@ def train(epoch):
             # Deactivated `TraceHook`, because it's computational intensive.
             # TraceHook(file_writer, FLAGS.log_frequency * 5),
             # LoggingHook.
-            tf_contrib.LoggerHook(loss)
+            LoggerHook(loss)
         ]
 
         # The MonitoredTrainingSession takes care of session initialization, session resumption,
